MapObjLib.cpp progress

+++ include/PowerPC_EABI_Support/Msl/MSL_C/MSL_Common/math.h
@@ -81,7 +81,7 @@ namespace std {
 inline float fabsf(float f) { return ::fabsf(f); }
 inline float abs(float f) { return ::fabs(f); }
 inline float fmodf(float x, float y) { return ::fmod(x, y); }
-inline float atan2f(float y, float x) { return ::atan2f(y, x); }
+inline float atan2f(float y, float x) { return ::atan2(y, x); }
 inline float sinf(float x) { return ::sin(x); }
 inline float cosf(float x) { return ::cos(x); }
 inline float tanf(float x) { return ::tan(x); }
